Create Support-less_Cyclotron_N-Filter_Washers
Full Cyclotron that does not need supports to print. This version also has the Washers and N-Filter removed so that they can be sanded and painted separately.  This version also has the N-Filter versions that have the welds added by Ricky Blakely. He has done an amazing job with this modification. Thank you Ricky. I have also included the N-Filters that can be used with the Full Clean Version.